refactor(hub): move record-keys → kernel/enclave
