+++ packages/hub/__tests__/extractpartition-blobs.test.ts
@@ -16,7 +16,7 @@ import { ref } from '../src/refs.js'
 import { ConflictError } from '../src/errors.js'
 import type { NoydbStore, EncryptedEnvelope, VaultSnapshot, KeyringFile } from '../src/types.js'
 import { decrypt } from '../src/kernel/enclave/crypto.js'
-import { unwrapCek } from '../src/record-keys/index.js'
+import { unwrapCek } from '../src/kernel/enclave/record-keys/index.js'
 import { withBlobs } from '../src/with-shape/blobs/index.js'
 import { BLOB_INDEX_COLLECTION, BLOB_CHUNKS_COLLECTION } from '../src/with-shape/blobs/blob-set.js'
 import { extractPartition } from '../src/with-cargo/extract-partition.js'
