+++ packages/hub/src/kernel/enclave/record-keys/deterministic.ts
@@ -16,8 +16,8 @@
  *
  * Internal subsystem — not exported as a `@noy-db/hub/*` subpath.
  */
-import { encryptDeterministic } from '../kernel/enclave/crypto.js'
-import type { NoydbStore } from '../types.js'
+import { encryptDeterministic } from '../crypto.js'
+import type { NoydbStore } from '../../../types.js'
 import type { RecordCodec } from './record-codec.js'
 
 /** Everything the moving deterministic-index methods touched on `this.*`. */
