+++ packages/hub/src/kernel/enclave/record-keys/lifecycle.ts
@@ -19,9 +19,9 @@
  * the collection (its lifetime is tied to `load()`), so callers cache the
  * returned key themselves.
  */
-import { encrypt, decrypt, generateDEK, wrapCek, unwrapCek } from '../kernel/enclave/crypto.js'
-import type { EncryptedEnvelope } from '../types.js'
-import type { Lru } from '../cache/index.js'
+import { encrypt, decrypt, generateDEK, wrapCek, unwrapCek } from '../crypto.js'
+import type { EncryptedEnvelope } from '../../../types.js'
+import type { Lru } from '../../../cache/index.js'
 
 /** Dependencies {@link resolveStableCek} needs from its collection. */
 export interface StableCekDeps {
