+++ packages/hub/src/kernel/enclave/record-keys/record-codec.ts
@@ -18,15 +18,15 @@
  *
  * Internal subsystem — not exported as a `@noy-db/hub/*` subpath.
  */
-import { encrypt, decrypt, encryptDeterministic, wrapCek, unwrapCek, deriveSealedFieldKey, deriveSealedFieldKeyFromCek } from '../kernel/enclave/crypto.js'
-import { NOYDB_FORMAT_VERSION, SealedHandle, type EncryptedEnvelope } from '../types.js'
+import { encrypt, decrypt, encryptDeterministic, wrapCek, unwrapCek, deriveSealedFieldKey, deriveSealedFieldKeyFromCek } from '../crypto.js'
+import { NOYDB_FORMAT_VERSION, SealedHandle, type EncryptedEnvelope } from '../../../types.js'
 import { isTombstone } from './tombstone.js'
 import { parseSealedSlot, dualReadSealedSlot } from './sealed-slot.js'
-import { DebugReservedFieldError } from '../errors.js'
-import { validateSchemaOutput, type StandardSchemaV1 } from '../schema.js'
-import type { CrdtMode, CrdtState } from '../with-commit/crdt/crdt.js'
-import type { CrdtStrategy } from '../with-commit/crdt/strategy.js'
-import type { Lru } from '../cache/index.js'
+import { DebugReservedFieldError } from '../../../errors.js'
+import { validateSchemaOutput, type StandardSchemaV1 } from '../../../schema.js'
+import type { CrdtMode, CrdtState } from '../../../with-commit/crdt/crdt.js'
+import type { CrdtStrategy } from '../../../with-commit/crdt/strategy.js'
+import type { Lru } from '../../../cache/index.js'
 
 /** Everything the moving crypto methods touched on `this.*`, as a flat context. */
 export interface RecordCodecContext<T> {
