+++ packages/hub/src/kernel/enclave/record-keys/sealing.ts
@@ -14,12 +14,12 @@
  * `sealed-record/`, so the host-side subpath stays DEK-free — they import only
  * the wire *types* from there.
  */
-import { encrypt, decrypt, generateDEK, wrapCek, unwrapCek, bufferToBase64, deriveSealedFieldKeyFromCek } from '../kernel/enclave/crypto.js'
-import { NOYDB_FORMAT_VERSION, type EncryptedEnvelope, type NoydbStore } from '../types.js'
+import { encrypt, decrypt, generateDEK, wrapCek, unwrapCek, bufferToBase64, deriveSealedFieldKeyFromCek } from '../crypto.js'
+import { NOYDB_FORMAT_VERSION, type EncryptedEnvelope, type NoydbStore } from '../../../types.js'
 import { dualReadSealedSlot } from './sealed-slot.js'
-import { RecordCekNotFoundError, ValidationError } from '../errors.js'
-import type { RecipientSealer } from '../with-party/team/managed-passphrase.js'
-import type { SealedCekBinding, SealedCekDeliveryEnvelope } from '../with-audit/sealed-record/types.js'
+import { RecordCekNotFoundError, ValidationError } from '../../../errors.js'
+import type { RecipientSealer } from '../../../with-party/team/managed-passphrase.js'
+import type { SealedCekBinding, SealedCekDeliveryEnvelope } from '../../../with-audit/sealed-record/types.js'
 
 const subtle = globalThis.crypto.subtle
