+++ packages/hub/src/with-audit/tiers/index.ts
@@ -22,8 +22,8 @@
  * Internal subsystem — not exported as a `@noy-db/hub/*` subpath.
  */
 import { encrypt, decrypt } from '../../kernel/enclave/crypto.js'
-import { unwrapCek, rewrapBodyToDek } from '../../record-keys/index.js'
-import type { RecordCodec } from '../../record-keys/record-codec.js'
+import { unwrapCek, rewrapBodyToDek } from '../../kernel/enclave/record-keys/index.js'
+import type { RecordCodec } from '../../kernel/enclave/record-keys/record-codec.js'
 import { TierDemoteDeniedError } from '../../errors.js'
 import { dekKey, assertTierAccess } from '../../with-party/team/tiers.js'
 import type { UnlockedKeyring } from '../../with-party/team/keyring.js'
