+++ packages/hub/src/with-cargo/extract-partition.ts
@@ -17,7 +17,7 @@ import {
   encryptBytesWithAAD,
   decryptBytesWithAAD,
 } from '../kernel/enclave/crypto.js'
-import { unwrapCek, wrapCek } from '../record-keys/index.js'
+import { unwrapCek, wrapCek } from '../kernel/enclave/record-keys/index.js'
 import {
   BLOB_COLLECTION,
   BLOB_INDEX_COLLECTION,
