+++ packages/hub/src/with-lookup/search/collection-facade.ts
@@ -24,7 +24,7 @@
  * Internal subsystem — not exported as a `@noy-db/hub/*` subpath.
  */
 import type { NoydbStore } from '../../types.js'
-import type { RecordCodec } from '../../record-keys/record-codec.js'
+import type { RecordCodec } from '../../kernel/enclave/record-keys/record-codec.js'
 import { stripI18nFilled, type I18nTextDescriptor } from '../../with-shape/i18n/core.js'
 import { isStaticDictDescriptor, type DictKeyDescriptor, type StaticDictDescriptor, type DictionaryHandle } from '../../with-shape/i18n/dictionary.js'
 import type { BlobSet } from '../../with-shape/blobs/blob-set.js'
